use context to pass schema
