+++ src/resources/Deployments/DeploymentCreate.js
@@ -1,4 +1,4 @@
-import { useEffect, useMemo, useState } from 'react';
+import { useContext, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import * as jp from 'jsonpath';
 import * as _ from 'lodash';
@@ -12,7 +12,7 @@ import {
   createDeploymentTemplate,
   createPresets,
 } from './templates';
-import { useGetSchema } from 'hooks/useGetSchema';
+import { SchemaContext } from 'shared/helpers/schema';
 
 const ISTIO_INJECTION_LABEL = 'sidecar.istio.io/inject';
 const ISTIO_INJECTION_ENABLED = 'true';
@@ -62,16 +62,7 @@ export default function DeploymentCreate({
     setCustomValid(hasAnyContainers);
   }, [deployment, setCustomValid]);
 
-  const resourceSchemaId = useMemo(
-    () => deployment?.apiVersion + '/' + deployment?.kind,
-    [], // eslint-disable-line react-hooks/exhaustive-deps
-  );
-  const { schema, loading, error } = useGetSchema({
-    schemaId: resourceSchemaId,
-  });
-  if (error) {
-    throw error;
-  }
+  const schema = useContext(SchemaContext);
 
   const handleNameChange = name => {
     jp.value(deployment, '$.metadata.name', name);
@@ -83,49 +74,44 @@ export default function DeploymentCreate({
   };
 
   return (
-    <>
-      {!loading ? (
-        <ResourceForm
-          {...props}
-          pluralKind="deployments"
-          singularName={t(`deployments.name_singular`)}
-          resource={deployment}
-          setResource={setDeployment}
-          onChange={onChange}
-          formElementRef={formElementRef}
-          presets={!initialUnchangedResource && createPresets(namespace, t)}
-          onPresetSelected={value => {
-            setDeployment(value.deployment);
+    <ResourceForm
+      {...props}
+      pluralKind="deployments"
+      singularName={t(`deployments.name_singular`)}
+      resource={deployment}
+      setResource={setDeployment}
+      onChange={onChange}
+      formElementRef={formElementRef}
+      presets={!initialUnchangedResource && createPresets(namespace, t)}
+      onPresetSelected={value => {
+        setDeployment(value.deployment);
+      }}
+      // create modal on a namespace details doesn't have the resourceUrl
+      createUrl={resourceUrl}
+      initialResource={initialResource}
+      initialUnchangedResource={initialUnchangedResource}
+      handleNameChange={handleNameChange}
+    >
+      {isIstioFeatureOn ? (
+        <ResourceForm.FormField
+          label={t('namespaces.create-modal.enable-sidecar')}
+          input={Inputs.Switch}
+          checked={isSidecarEnabled}
+          onChange={() => {
+            setSidecarEnabled(value => !value);
+            setIsChanged(true);
           }}
-          // create modal on a namespace details doesn't have the resourceUrl
-          createUrl={resourceUrl}
-          initialResource={initialResource}
-          initialUnchangedResource={initialUnchangedResource}
-          handleNameChange={handleNameChange}
-          schema={schema}
-        >
-          {isIstioFeatureOn ? (
-            <ResourceForm.FormField
-              label={t('namespaces.create-modal.enable-sidecar')}
-              input={Inputs.Switch}
-              checked={isSidecarEnabled}
-              onChange={() => {
-                setSidecarEnabled(value => !value);
-                setIsChanged(true);
-              }}
-            />
-          ) : null}
-
-          <AdvancedContainersView
-            resource={deployment}
-            setResource={setDeployment}
-            onChange={onChange}
-            namespace={namespace}
-            createContainerTemplate={createContainerTemplate}
-            schema={schema}
-          />
-        </ResourceForm>
+        />
       ) : null}
-    </>
+
+      <AdvancedContainersView
+        resource={deployment}
+        setResource={setDeployment}
+        onChange={onChange}
+        namespace={namespace}
+        createContainerTemplate={createContainerTemplate}
+        schema={schema}
+      />
+    </ResourceForm>
   );
 }
