+++ src/resources/Jobs/JobCreate.js
@@ -1,4 +1,4 @@
-import { useEffect, useMemo, useState } from 'react';
+import { useContext, useEffect, useState } from 'react';
 import * as jp from 'jsonpath';
 import { useTranslation } from 'react-i18next';
 
@@ -10,7 +10,7 @@ import { createJobPresets, createJobTemplate } from './templates';
 import { JobSpecSection } from './SpecSection';
 import { ContainersSection } from './ContainersSection';
 import { MessageStrip } from '@ui5/webcomponents-react';
-import { useGetSchema } from 'hooks/useGetSchema';
+import { SchemaContext } from 'shared/helpers/schema';
 
 function isJobValid(job = {}) {
   const isNameValid = jp.value(job, '$.metadata.name');
@@ -52,50 +52,37 @@ export default function JobCreate({
     setCustomValid(isJobValid(job));
   }, [job, setCustomValid]);
 
-  const resourceSchemaId = useMemo(
-    () => job?.apiVersion + '/' + job?.kind,
-    [], // eslint-disable-line react-hooks/exhaustive-deps
-  );
-  const { schema, loading, error } = useGetSchema({
-    schemaId: resourceSchemaId,
-  });
-  if (error) {
-    throw error;
-  }
+  const schema = useContext(SchemaContext);
 
   return (
-    <>
-      {!loading ? (
-        <ResourceForm
-          {...props}
-          pluralKind="jobs"
-          singularName={t(`jobs.name_singular`)}
-          resource={job}
-          setResource={setJob}
-          initialResource={initialResource}
-          initialUnchangedResource={initialUnchangedResource}
-          onChange={onChange}
-          formElementRef={formElementRef}
-          presets={
-            !initialUnchangedResource &&
-            createJobPresets(namespace, t, defaultSidecarAnnotations)
-          }
-          createUrl={resourceUrl}
-          schema={schema}
-        >
-          <JobSpecSection
-            propertyPath="$.spec"
-            readOnly={!!initialUnchangedResource}
-          />
-          <ContainersSection
-            propertyPath="$.spec.template.spec.containers"
-            readOnly={!!initialUnchangedResource}
-          />
-          <MessageStrip design="Information" hideCloseButton>
-            {t('jobs.create-modal.containers-readonly-in-edit')}
-          </MessageStrip>
-        </ResourceForm>
-      ) : null}
-    </>
+    <ResourceForm
+      {...props}
+      pluralKind="jobs"
+      singularName={t(`jobs.name_singular`)}
+      resource={job}
+      setResource={setJob}
+      initialResource={initialResource}
+      initialUnchangedResource={initialUnchangedResource}
+      onChange={onChange}
+      formElementRef={formElementRef}
+      presets={
+        !initialUnchangedResource &&
+        createJobPresets(namespace, t, defaultSidecarAnnotations)
+      }
+      createUrl={resourceUrl}
+      schema={schema}
+    >
+      <JobSpecSection
+        propertyPath="$.spec"
+        readOnly={!!initialUnchangedResource}
+      />
+      <ContainersSection
+        propertyPath="$.spec.template.spec.containers"
+        readOnly={!!initialUnchangedResource}
+      />
+      <MessageStrip design="Information" hideCloseButton>
+        {t('jobs.create-modal.containers-readonly-in-edit')}
+      </MessageStrip>
+    </ResourceForm>
   );
 }
