+++ src/resources/createResourceRoutes.js
@@ -1,7 +1,7 @@
 import React, { Suspense, useEffect, useMemo } from 'react';
 
 import { useRecoilState } from 'recoil';
-import { Route, useSearchParams, useParams } from 'react-router-dom';
+import { Route, useParams, useSearchParams } from 'react-router-dom';
 import pluralize from 'pluralize';
 import { FlexibleColumnLayout } from '@ui5/webcomponents-react';
 import { useTranslation } from 'react-i18next';
@@ -15,8 +15,10 @@ import {
   usePrepareListProps,
 } from './helpers';
 import { ErrorBoundary } from 'shared/components/ErrorBoundary/ErrorBoundary';
-import { ResourceCreate } from 'shared/components/ResourceCreate/ResourceCreate';
 import { useUrl } from 'hooks/useUrl';
+import { useGetSchema } from 'hooks/useGetSchema';
+import { SchemaContext } from 'shared/helpers/schema';
+import { ResourceCreate } from 'shared/components/ResourceCreate/ResourceCreate';
 
 export const createPath = (
   config = { detailsView: false, pathSegment: '' },
@@ -147,26 +149,49 @@ const ColumnWrapper = ({
     detailsMidColumn = detailsComponent;
   }
 
+  const resourceSchemaId = useMemo(
+    //In each index.js resourceType is in plural form. I don't know if it's correct
+    () =>
+      [
+        props?.apiGroup,
+        props?.apiVersion,
+        props?.resourceType.slice(0, -1),
+      ].join('/'),
+    [], // eslint-disable-line react-hooks/exhaustive-deps
+  );
+  const { schema, loading, error } = useGetSchema({
+    schemaId: resourceSchemaId,
+  });
+  if (error) {
+    //TODO:
+    // throw error;
+  }
+  if (loading) {
+    return null;
+  }
+
   const createMidColumn = (
-    <ResourceCreate
-      title={elementCreateProps.resourceTitle}
-      confirmText={t('common.buttons.create')}
-      layoutCloseCreateUrl={layoutCloseCreateUrl}
-      renderForm={renderProps => {
-        const createComponent =
-          create &&
-          create?.type !== null &&
-          (layoutState?.showCreate?.resourceType || props?.resourceType) &&
-          React.cloneElement(create, {
-            ...elementCreateProps,
-            ...renderProps,
-            enableColumnLayout: true,
-            layoutNumber: 'StartColumn',
-            resource: layoutState?.showCreate?.resource, // For ResourceCreate we want to set layoutNumber to previous column so detail are opened instead of create
-          });
-        return <ErrorBoundary>{createComponent}</ErrorBoundary>;
-      }}
-    />
+    <SchemaContext.Provider value={schema}>
+      <ResourceCreate
+        title={elementCreateProps.resourceTitle}
+        confirmText={t('common.buttons.create')}
+        layoutCloseCreateUrl={layoutCloseCreateUrl}
+        renderForm={renderProps => {
+          const createComponent =
+            create &&
+            create?.type !== null &&
+            (layoutState?.showCreate?.resourceType || props?.resourceType) &&
+            React.cloneElement(create, {
+              ...elementCreateProps,
+              ...renderProps,
+              enableColumnLayout: true,
+              layoutNumber: 'StartColumn',
+              resource: layoutState?.showCreate?.resource, // For ResourceCreate we want to set layoutNumber to previous column so detail are opened instead of create
+            });
+          return <ErrorBoundary>{createComponent}</ErrorBoundary>;
+        }}
+      />
+    </SchemaContext.Provider>
   );
 
   return (
